refactor: 💡 filter, guard 위치를 common 밑으로 변경
해당 필터와 가드는 도메인과 상관 없이 여러 곳에서 쓰일 것이기에 common 디렉터리 밑으로 변경했습니다.
