+++ backend/src/common/guard/connect.guard.ts
@@ -2,7 +2,7 @@ import { CanActivate, ExecutionContext, Injectable } from '@nestjs/common';
 import { WsException } from '@nestjs/websockets';
 
 @Injectable()
-export class ConnectionGuard implements CanActivate {
+export class ConnectGuard implements CanActivate {
 
   canActivate(context: ExecutionContext): boolean {
     const client = context.switchToWs().getClient();
