+++ backend/src/common/guard/host.guard.ts
@@ -1,10 +1,11 @@
 import { CanActivate, ExecutionContext, Injectable } from '@nestjs/common';
-import { RoomsService } from '../service/rooms.service';
+import { RoomsService } from '../../rooms/service/rooms.service';
 import { WsException } from '@nestjs/websockets';
 
 @Injectable()
 export class HostGuard implements CanActivate {
-  constructor(private readonly roomsService: RoomsService) {}
+  constructor(private readonly roomsService: RoomsService) {
+  }
 
   canActivate(context: ExecutionContext): boolean {
     const client = context.switchToWs().getClient();
